move year ts page into different dictionary
