+++ src/routes/+page.ts
@@ -1,6 +1,6 @@
 import type { LoadYear } from 'types/loadTypes';
 
-import { loadYearAsync } from './year/[year]/year';
+import { loadYearAsync } from 'pageHelper/loadYear';
 
 export async function load({ fetch }: {
     fetch: typeof globalThis.fetch
